+++ modules/fs/init_fs.sh
@@ -1,12 +1,12 @@
 init_fs () {
     folder=$1
 
-    print "Initializing file system at $folder..."
-
     mkdir -p $folder/services
     mkdir -p $folder/logs
     
     mkdir -p $folder/config
     mkdir -p $folder/config/nginx
     mkdir -p $folder/config/supervisor
+
+    print "${SYMBOL_OK} File system initialized at $folder..."
 }