+++ modules/services/install_services.sh
@@ -2,22 +2,16 @@ install_services () {
     print "Installing missing services..."
     
     if ! $(is_nginx_installed) -eq "true"; then
-        if $(prompt_installation_yn "Nginx") -eq "true"; then
-            install_nginx
-        fi
+        install_nginx
     fi
 
     if ! $(is_supervisor_installed) -eq "true"; then
-        if $(prompt_installation_yn "Supervisor") -eq "true"; then
-            install_supervisor
-        fi
+        install_supervisor
     fi
 
     if ! $(is_git_installed) -eq "true"; then
-        if $(prompt_installation_yn "Git") -eq "true"; then
-            install_git
-        fi
+        install_git
     fi
 
-    print "Services installed."
+    print "${SYMBOL_OK} All services has been installed."
 }