+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 
 ### Added
+- [2024-04-28] Add pretty print to the installations and init_fs processes
 - [2024-04-28] Add filesystem initialization for vm folders
 - [2024-04-28] Add build script that bundles the file into build/index.sh